requirements.txt
selenium
psycopg2-binary
python-dotenv
tqdm
fastapi
uvicorn[standard]
requests